requirements.txt
django==3.2.7
psycopg2-binary==2.9.1
django-countries==7.2.1
pyuca==1.2
django-mptt==0.13.3
Pillow==8.3.2
sorl-thumbnail==12.7.0
django-storages==1.11.1
boto3==1.18.47
dj-database-url==0.5.0
whitenoise==5.3.0
celery==5.1.2
python-memcached==1.59
gunicorn==20.1.0
django-ses==2.3.0
django-storages==1.11.1
django-braces==1.14.0
